LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/cookies/partitioned-cookies/partitioned-cookies-a-b-a-embed.tentative.https-expected.txt

CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testharness.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/cookies/resources/testharness-helpers.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/cookies/resources/cookie-helper.sub.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/common/get-host-info.sub.js from asking for credentials because it is a cross-origin request.


PASS Setting Partitioned cookie on top-level site and embedding a cross-site iframe
PASS Embed same-site embed in cross-site
FAIL Same-site embed with a cross-site parent partitioned cookie access assert_false: expected false got true
